[FR-56] feat: 순환 참조를 해결하기 위해 RedisRepository Component 추가. AuthService에서 RedisRepository Bean을 참조하도록 재구성.
